requirements.txt
python-telegram-bot==20.7
gspread==5.12.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
python-dotenv==1.0.0
APScheduler==3.10.4
pytz==2023.3
